Combat 2d: Damage is entity-native
The damage pipeline, 1353 lines, is CombatSystem::Damage(victim, attacker, dam,
type) and CHARACTER::Damage is gone. The receiver used to be the thing being
hit, so every bare accessor in the body was a read on the victim; that is what
the entity in each of them now is. Sixty-five distinct member calls became
entity calls, and the arithmetic, the branch order and the flag handling are
untouched.

Two more fields moved rather than being resolved through a character on every
touch. cannot_dead, the GM armada flag, is ecs::UndyingState. m_dwKillerPID,
which the damage writes and the death reads, is ecs::KillerPID.

One thing is still asked of a character, once, at the top and named: the damage
map that decides drops and experience is a CHARACTER member and its own
migration, so the victim is resolved for it and for nothing else. The three
other legacy calls inside the reference-damage helpers - GetArrowAndBow,
GetSoulItemDamage - are named where they happen.

This took three attempts at the mechanical part and both failures are worth
recording. The first placed the converted body where the old forwarder sat,
which is above the file-local statics it calls. The second renamed
m_map_kDamage across the whole file rather than inside the function, breaking
five unrelated CHARACTER methods that legitimately use it. Both were caught by
the compiler rather than shipped, but both come from the same habit: applying a
rewrite to a file when it belongs to a span. The locator now refuses anything
that is not a definition, and the rename is bounded to the extracted body.

Release and RelWithDebInfo clean, 15/15 ctest, CombatStateTests green under
AddressSanitizer, server boots with no crash and no [ECS] report.

+++ tests/server/AffectLifecycleTests.cpp
@@ -824,7 +824,7 @@ void ecs::MovementSystem::SyncDestinationClear(entt::entity e) {
 void ecs::PlayerRuntime::SetPosition(entt::entity e, int position) {
     g_registry.get<ecs::CharacterRuntimeFlagsComponent>(e).position = position;
 }
-bool CombatSystem::Damage(entt::entity,entt::entity,int,unsigned char) { UnexpectedService(__func__); }
+bool CombatSystem::Damage(entt::entity,entt::entity,int64_t,unsigned char) { UnexpectedService(__func__); }
 void CombatSystem::SetComboSequence(entt::entity,unsigned char) { UnexpectedService(__func__); }
 void CombatSystem::SetValidComboInterval(entt::entity,int) { UnexpectedService(__func__); }
 unsigned int CGuild::UnderAnyWar(unsigned char) { UnexpectedService(__func__); }
